main.cpp: allow configuration of server endpoint
Adds support for a simple configuration file that just contains
a server address and a port in a file placed at
/sdcard/BMBFData/BeatTogether.cfg

Only one line is read from the file and it should be in the following
format:
<hostname> <port>

This allows overwriting the target server for testing or if
running an own server. If the file is not present, the default will
be used instead.